+++ CHANGELOG.rst
@@ -2,11 +2,12 @@
 Change Log
 ==========
 
-HEAD
-----
+1.9.2
+-----
 * Added ``EWSTimeZone.localzone()`` to get the local timezone
 * Support ``some_folder.get(item_id=..., changekey=...)`` as a shortcut to get a single item when you know the ID and
   changekey.
+* Support attachments on Exchange 2007
 
 1.9.1
 -----